refactor: 조회수 증가 로직 리펙토링 -> DB사용X, ehcache + 쿠키 기반 중복 체크, 낙관적 락 적용(동시성 문제 해소)

+++ noteJ/src/main/java/com/github/jkky_98/noteJ/service/WriteServiceLocal.java
@@ -187,7 +187,9 @@ public class WriteServiceLocal implements WriteService {
         String oldThumbnail = post.getThumbnail(); // ✅ 기존 썸네일 저장
 
         post.updateThumbnail(newThumbnail);
-        deleteOldThumbnail(oldThumbnail);
+        if (!oldThumbnail.equals(DEFAULT_POST_PIC)) {
+            deleteOldThumbnail(oldThumbnail);
+        }
     }
 
     /**
