+++ noteJ/src/main/java/com/github/jkky_98/noteJ/service/WriteServiceProd.java
@@ -240,7 +240,9 @@ public class WriteServiceProd implements WriteService{
         String oldThumbnail = post.getThumbnail(); // ✅ 기존 썸네일 저장
 
         post.updateThumbnail(newThumbnail);
-        deleteOldThumbnail(oldThumbnail);
+        if (!oldThumbnail.equals(DEFAULT_POST_PIC)) {
+            deleteOldThumbnail(oldThumbnail);
+        }
     }
 
     /**
